remove old authentication hook
